Created the Inital Profile
Enabel the user to create a profile on first entry.
Details are stored in storage and persistent throughout sessions.
User can add their name, profile image and select a theme.
This will dictate the overall look of the app.
The settings includes and empty section array.
This is where we will add the Sections which will be Objects
storing the information the app will render the different sections from.